Restored work of hamburger menu
