+++ src/styles/blocks/header/header__title.css
@@ -7,6 +7,7 @@
 
 @media screen and (max-width : 320px) {
   .header__title {
-    font-size: 20px;
+    font-size: 16px;
+    line-height: 24px;
   }
 }
